feature-/U-020 add user signIn and signUp function

+++ app/src/main/java/com/swsnack/catchhouse/data/userdata/api/AppAPIManager.java
@@ -7,7 +7,9 @@ import com.facebook.AccessToken;
 import com.facebook.GraphRequest;
 import com.google.android.gms.tasks.OnFailureListener;
 import com.google.android.gms.tasks.OnSuccessListener;
+import com.google.firebase.FirebaseException;
 import com.google.firebase.auth.AuthCredential;
+import com.google.firebase.auth.AuthResult;
 import com.google.firebase.auth.FirebaseAuth;
 import com.swsnack.catchhouse.constants.Constants;
 import com.swsnack.catchhouse.data.userdata.APIManager;
@@ -26,14 +28,14 @@ public class AppAPIManager implements APIManager {
     }
 
     @Override
-    public void firebaseSignUp(@NonNull AuthCredential authCredential, @NonNull OnSuccessListener onSuccessListener, @NonNull OnFailureListener onFailureListener) {
+    public void firebaseSignUp(@NonNull AuthCredential authCredential, @NonNull OnSuccessListener<AuthResult> onSuccessListener, @NonNull OnFailureListener onFailureListener) {
         FirebaseAuth.getInstance().signInWithCredential(authCredential)
                 .addOnSuccessListener(onSuccessListener)
                 .addOnFailureListener(onFailureListener);
     }
 
     @Override
-    public void firebaseSignUp(@NonNull String email, @NonNull String password, @NonNull OnSuccessListener onSuccessListener, @NonNull OnFailureListener onFailureListener) {
+    public void firebaseSignUp(@NonNull String email, @NonNull String password, @NonNull OnSuccessListener<AuthResult> onSuccessListener, @NonNull OnFailureListener onFailureListener) {
         FirebaseAuth.getInstance().createUserWithEmailAndPassword(email, password)
                 .addOnSuccessListener(onSuccessListener)
                 .addOnFailureListener(onFailureListener);
@@ -49,14 +51,18 @@ public class AppAPIManager implements APIManager {
     }
 
     @Override
-    public void firebaseSignIn(@NonNull String email, @NonNull String password, @NonNull OnSuccessListener onSuccessListener, @NonNull OnFailureListener onFailureListener) {
+    public void firebaseSignIn(@NonNull String email, @NonNull String password, @NonNull OnSuccessListener<AuthResult> onSuccessListener, @NonNull OnFailureListener onFailureListener) {
         FirebaseAuth.getInstance().signInWithEmailAndPassword(email, password)
                 .addOnSuccessListener(onSuccessListener)
                 .addOnFailureListener(onFailureListener);
     }
 
     @Override
-    public void firebaseDeleteUser(@NonNull OnSuccessListener onSuccessListener, @NonNull OnFailureListener onFailureListener) {
+    public void firebaseDeleteUser(@NonNull OnSuccessListener<Void> onSuccessListener, @NonNull OnFailureListener onFailureListener) {
+        if (FirebaseAuth.getInstance().getCurrentUser() == null) {
+            onFailureListener.onFailure(new FirebaseException("user is not signIn"));
+            return;
+        }
         FirebaseAuth.getInstance().getCurrentUser()
                 .delete()
                 .addOnSuccessListener(onSuccessListener)
